requirement.txt
llama-index==0.12.5
pypdf
transformers==4.46.3
einops 
accelerate==1.3.0
langchain==0.3.12
llama-index-llms-huggingface==0.4.1
bitsandbytes==0.45.0
langchain-community==0.3.12
llama-index-embeddings-langchain==0.3.0
sentence-transformers
llama-index-llms-groq

# server
fastapi 
uvicorn 
streamlit 
requests 
python-multipart
python-dotenv